Move imports: from manticore import issymbolic, istainted
