+++ examples/script/introduce_symbolic_bytes.py
@@ -1,7 +1,9 @@
 #!/usr/bin/env python
 
 import sys
-from manticore.native import Manticore, issymbolic
+
+from manticore import issymbolic
+from manticore.native import Manticore
 
 '''
 Replaces a variable that controls program flow with a tainted symbolic value. This
