+++ manticore/native/__init__.py
@@ -4,6 +4,5 @@ install_helper.ensure_native_deps()
 
 # Exports (for `from manticore.native import ...`)
 from .manticore import Manticore
-from ..utils.helpers import issymbolic
 from .models import variadic
 from . import cpu
